server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "API de HabitFlow"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "email-validator>=2.3.0",
    "fastapi[standard] (>=0.119.1,<0.120.0)",
    "pydantic>=2.12.3",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "python-decouple>=3.8",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]
